+++ packages/react/src/index.ts
@@ -31,7 +31,21 @@ export { default as SignOutButton } from './SignOutButton';
 export { default as SignedIn } from './SignedIn';
 
 export { default as SignedOut } from './SignedOut';
-
+export type {
+  AuthenticatorSetup,
+  EmailSetup,
+  EnrollmentError,
+  EnrollMfaProps,
+  ChallengeRedirect,
+  MfaMethod,
+  MfaSetup,
+} from './MultiFactor';
+export {
+  useMultiFactor,
+  MultiFactorAuthenticator,
+  MultiFactorEmail,
+  MultiFactorChallenge,
+} from './MultiFactor';
 export {
   default as TenantSelector,
   useTenantId,